+++ merkle/util.go
@@ -1,7 +1,7 @@
 package merkle
 
 // return the number of 1 bit
-func countBit(num uint32) uint {
+func countBit(num uint64) uint {
 	var count uint
 	for num != 0 {
 		num &= (num - 1)
@@ -10,13 +10,13 @@ func countBit(num uint32) uint {
 	return count
 }
 
-func isPower2(num uint32) bool {
+func isPower2(num uint64) bool {
 	return countBit(num) == 1
 }
 
 // return the position of the heightest 1 bit
 // 1-based index
-func highBit(num uint32) uint {
+func highBit(num uint64) uint {
 	var hiBit uint
 	for num != 0 {
 		num >>= 1
@@ -27,6 +27,6 @@ func highBit(num uint32) uint {
 
 // return the position of the lowest 1 bit
 // 1-based index
-func lowBit(num uint32) uint {
+func lowBit(num uint64) uint {
 	return highBit(num & -num)
 }
